TI-S2-revise-D03

+++ src/main/java/acme/entities/enrolments/Enrolment.java
@@ -43,7 +43,7 @@ public class Enrolment extends AbstractEntity {
 	protected String			goals;
 
 	@Transient
-	protected Integer			workTime			= 0;
+	protected Double			workTime			= 0.0;
 
 	@NotNull
 	@Valid
